# docs/source/API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reachy Mini REST API",
    "version": "0.1.0",
    "description": "HTTP and WebSocket API exposed by the Reachy Mini daemon. Interactive Swagger docs are available at `/docs` when the daemon is running."
  },
  "paths": {
...
    "/api/volume/set": {
      "post": {
        "summary": "Set Volume",
        "description": "Set the output volume level and play a test sound.",
        "operationId": "set_volume_api_volume_set_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VolumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/volume/microphone/set": {
      "post": {
        "summary": "Set Microphone Volume",
        "description": "Set the microphone input volume level.",
        "operationId": "set_microphone_volume_api_volume_microphone_set_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VolumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VolumeRequest": {
        "properties": {
          "volume": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Volume",
            "description": "Volume level (0-100)"
          }
        },
        "type": "object",
        "required": [
          "volume"
        ],
        "title": "VolumeRequest",
        "description": "Request model for setting volume."
      },
      "VolumeResponse": {
        "properties": {
          "volume": {
            "type": "integer",
            "title": "Volume"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "device": {
            "type": "string",
            "title": "Device"
          }
        },
        "type": "object",
        "required": [
          "volume",
          "platform",
          "device"
        ],
        "title": "VolumeResponse",
        "description": "Response model for volume operations."
      },
...
    }
  }
}